backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.34
python-dotenv==1.0.1
psycopg[binary]==3.1.18
requests==2.32.3
pgvector==0.2.5
pydantic==2.8.2
numpy
h3==3.7.6
